+++ packages/awino-ui/src/components/buttons/ConnectButton.tsx
@@ -1,32 +1,39 @@
-import { useCallback } from 'react';
+import { useCallback, useState } from 'react';
 
 import { useTranslation } from 'next-i18next';
 
 import { Button } from '@mui/material';
 
-import { useAppDispatch, useAppSelector } from '@/app/hooks';
-import { connect, disconnect } from '@/app/state/slices/account';
+import { /* useAppDispatch, */ useAppSelector } from '@/app/hooks';
+// import { connect, disconnect } from '@/app/state/slices/account';
+
+import WalletConnect from '../modals/WalletConnect/WalletConnect';
 
 export default function ConnectButton(props) {
   const { t } = useTranslation();
-  const dispatch = useAppDispatch();
+  // const dispatch = useAppDispatch();
+  const [open, setOpen] = useState(false);
   const { connected } = useAppSelector((state) => ({
     connected: state.account.connected,
   }));
 
   /* TODO WIP Update when connection logic is defined  */
-  const handleConnectToggle = useCallback(() => {
-    if (connected) {
-      dispatch(disconnect());
-    } else {
-      dispatch(connect());
-    }
+  const handleOpen = useCallback(() => {
+    setOpen(true);
+    // if (connected) {
+    //   dispatch(disconnect());
+    // } else {
+    //   dispatch(connect());
+    // }
     // console.log('connecting/disconnecting');
-  }, [connected, dispatch]);
+  }, [setOpen /* connected, dispatch */]);
 
   return (
-    <Button color="primary" onClick={handleConnectToggle} {...props}>
-      {t(`account.${connected ? 'dis' : ''}connect`)}
-    </Button>
+    <>
+      <Button color="primary" onClick={handleOpen} {...props}>
+        {t(`account.${connected ? 'dis' : ''}connect`)}
+      </Button>
+      {open && <WalletConnect open={open} setOpen={setOpen} />}
+    </>
   );
 }