+++ packages/awino-ui/src/components/web3/ChainId.tsx
@@ -1,0 +1,12 @@
+import { useWeb3React } from '@web3-react/core';
+
+import LabelValue from '../general/LabelValue/LabelValue';
+
+const ChainId = () => {
+  const { chainId } = useWeb3React();
+
+  const value = chainId ?? '';
+  return <LabelValue id="chainId" value={value} labelProps={{ children: 'Chain Id' }} />;
+};
+
+export default ChainId;