+++ packages/awino-ui/src/pages/index.tsx
@@ -26,7 +26,7 @@ const IndexPage: NextPage = () => {
   return (
     <>
       <Head>
-        {/* eslint-disable-next-line @next/next/no-page-custom-font, @next/next/google-font-display */}
+        {/*  eslint-disable-next-line @next/next/no-page-custom-font, @next/next/google-font-display */}
         <link
           href="https://fonts.googleapis.com/css2?family=Poiret+One&family=MuseoModerno&display=block"
           rel="stylesheet"